Some path types, functions and proofs
Spent hours trying to prove some negations of equality types. Turned out
all I needed was to recognize that the Not type introduced an extra
argument to the theorem (the equality), match on that equality with
`Refl`, and then let Idris take it from there with `impossible`. Sometimes
I badly overcomplicate things. :'''( OTOH I'm glad when Idris can do
most of the work... doesn't always happen that way.